refactor: :recycle: refactor ProductItem & post-item to use image helper
