+++ client/src/helpers/image-helper.ts
@@ -1,3 +1,6 @@
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import imageCompression from 'browser-image-compression';
 
@@ -36,4 +39,53 @@ const saveImages = async (images: File[]) => {
 	return imgArr;
 };
 
-export { compressImage, saveImages };
+//Gets image hashes and info
+const fetchImageData = async (data, setHasImage, setImageData, setImageError) => {
+	try {
+		if (data) {
+			if (data.imageHash!.length > 0) {
+				setHasImage(true);
+				if (Array.isArray(data.imageHash)) {
+					await Promise.all(
+						data.imageHash.map(async (element: string) => {
+							await fetchImages(element, data.imageHash, setImageData, setImageError);
+						})
+					);
+				}
+			}
+		}
+	} catch (err) {
+		console.log(err);
+	}
+};
+
+//Gets images themselves
+const fetchImages = async (imageHash: string, imageHashes, setImageData, setImageError) => {
+	try {
+		const response = await fetch(`/api/upload?imageHash=${imageHash}`);
+		if (response.ok) {
+			const blob = await response.blob();
+			const imageUrl = URL.createObjectURL(blob);
+			//Below used to render images in the order they were uploaded to db...
+			// Find the index of the current imageHash in the imageHashes array
+			const index = imageHashes.indexOf(imageHash);
+			if (index !== -1) {
+				// Update the state at the corresponding index
+				setImageData((prevImageData) => {
+					const newData = [...prevImageData];
+					newData[index] = imageUrl;
+					return newData;
+				});
+			} else {
+				console.log('Image hash not found in imageHashes array');
+			}
+		} else {
+			console.log('Error fetching image');
+			setImageError(true);
+		}
+	} catch (err) {
+		console.log(err);
+	}
+};
+
+export { compressImage, saveImages, fetchImageData };
